+++ busesProject/src/main/java/busesProject/models/Autobus.java
@@ -1,6 +1,5 @@
 package busesProject.models;
 
-
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -17,9 +16,11 @@ public class Autobus {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_autobus")
-    private Integer idAutobus;
+    private Long idAutobus;
+
     @Column(nullable = false, unique = true, length = 20)
     private String placa;
+
     @ManyToOne
     @JoinColumn(name = "id_tipo",nullable = false)
     private TipoBus tipo;
