+++ busesProject/src/main/java/busesProject/models/Permiso.java
@@ -16,7 +16,7 @@ public class Permiso {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_permiso")
-    private Integer idPermiso;
+    private Long idPermiso;
 
     @Column(nullable = false, unique = true, length = 100)
     private String nombre;
