+++ busesProject/src/main/java/busesProject/models/Reserva.java
@@ -18,16 +18,20 @@ public class Reserva {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_reserva")
-    private Integer idReserva;
+    private Long idReserva;
+
     @ManyToOne
     @JoinColumn(name="id_usuario", nullable = false)
     private Usuario usuario;
+
     @Temporal(TemporalType.TIMESTAMP)
     @Column(name = "fecha_emision", nullable = false)
     private Date fechaEmision;
+
     @ManyToOne
     @JoinColumn(name="id_ciudad_origen", nullable = false)
     private Ciudad ciudadOrigen;
+
     @ManyToOne
     @JoinColumn(name="id_ciudad_destino", nullable = false)
     private Ciudad ciudadDestino;
