+++ busesProject/src/main/java/busesProject/models/TipoBus.java
@@ -16,13 +16,17 @@ public class TipoBus {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_tipo")
-    private Integer idTipo;
+    private Long idTipo;
+
     @Column(nullable = false, unique = true, length = 50)
     private String nombre;
+
     @Column(name = "capacidad_maxima",nullable = false)
     private Integer capacidadMaxima;
+
     @Column(nullable = false)
     private Integer filas;
+
     @Column(name = "asientos_por_fila",nullable = false)
     private Integer asientosPorFila;
 }
